refactoring positioning services
