+++ app/com/yetu/oauth2provider/services/data/interface/IAuthCodeAccessTokenService.scala
@@ -1,4 +1,4 @@
-package com.yetu.oauth2provider.services.data.iface
+package com.yetu.oauth2provider.services.data.interface
 
 import com.yetu.oauth2provider.oauth2.models.YetuUser
 
@@ -18,11 +18,13 @@ trait IAuthCodeAccessTokenService {
     val saveAuthInfoToken = authInfo.clientId.map(
       clientId => saveAuthInfoToAccessToken(authInfo.user.identityId.userId + clientId, accessToken))
 
-    for {
+    val result = for {
       token <- saveToken
       info <- saveAuthInfo
-      client <- saveAuthInfoToken.getOrElse(Future.successful())
+      client <- saveAuthInfoToken.getOrElse(Future.successful(Unit))
     } yield client
+
+    result.map(_ => Unit)
   }
 
   def saveAuthCode(code: String, authInfo: AuthInfo[YetuUser]): Future[Unit]
