+++ app/com/yetu/oauth2provider/services/data/interface/IClientService.scala
@@ -1,9 +1,9 @@
 package com.yetu.oauth2provider
 package services
 package data
-package iface
+package interface
 
-import com.yetu.oauth2provider.oauth2.models.{ OAuth2Client, ClientPermission }
+import com.yetu.oauth2provider.oauth2.models.OAuth2Client
 
 trait IClientService {
   def saveClient(client: OAuth2Client, ignoreEntryAlreadyExists: Boolean = false): Unit
@@ -13,7 +13,4 @@ trait IClientService {
   def deleteClient(client: OAuth2Client)
 
   def deleteClient(clientId: String)
-
-  def deleteAllClients(): Unit //do nothing, can be overridden for tests.
-
 }
