+++ app/com/yetu/oauth2provider/services/data/memory/MemoryClientService.scala
@@ -1,8 +1,7 @@
-package com.yetu.oauth2provider.services.data
+package com.yetu.oauth2provider.services.data.memory
 
-import com.yetu.oauth2provider.oauth2.models.{ ClientPermission, OAuth2Client }
-import com.yetu.oauth2provider.services.data.iface.IClientService
-import play.api.Logger
+import com.yetu.oauth2provider.oauth2.models.OAuth2Client
+import com.yetu.oauth2provider.services.data.interface.IClientService
 
 class MemoryClientService extends IClientService {
 
@@ -13,13 +12,7 @@ class MemoryClientService extends IClientService {
   }
 
   override def findClient(clientId: String): Option[OAuth2Client] = {
-
     clients.get(clientId)
-  }
-
-  override def deleteAllClients(): Unit = {
-    Logger.warn("delete All Clients was executed")
-    clients = Map[String, OAuth2Client]()
   }
 
   override def deleteClient(client: OAuth2Client) = clients -= client.clientId
@@ -29,6 +22,5 @@ class MemoryClientService extends IClientService {
 }
 
 object MemoryClientService {
-
   var clients = Map[String, OAuth2Client]()
 }
