+++ app/com/yetu/oauth2provider/services/data/memory/MemoryUserService.scala
@@ -1,7 +1,7 @@
-package com.yetu.oauth2provider.services.data
+package com.yetu.oauth2provider.services.data.memory
 
 import com.yetu.oauth2provider.oauth2.models.YetuUser
-import com.yetu.oauth2provider.services.data.iface.IMailTokenService
+import com.yetu.oauth2provider.services.data.interface.IMailTokenService
 import securesocial.core.providers.MailToken
 import securesocial.core.services.UserService
