+++ test/com/yetu/oauth2provider/base/BaseMethods.scala
@@ -1,7 +1,7 @@
 package com.yetu.oauth2provider.base
 
 import com.yetu.oauth2provider.oauth2.models.YetuUser
-import com.yetu.oauth2provider.services.data.iface.IAuthCodeAccessTokenService
+import com.yetu.oauth2provider.services.data.interface.IAuthCodeAccessTokenService
 import com.yetu.oauth2provider.utils.Config
 import play.api.Logger
